EOI table
hi Guys, 
I have created a table for you, EOI, please have a look.
i Will love to mention few things in this. 
here is the summary dwn below :- 
 The EOInumber is set a primary key which is main key at the movement, it is following auto increment that the id will be created own.
when the new id is created the status is created a new. 
i have basically provided    5 skills and one for other. 
i basically think we  can use the sort function of the MySQL to sort the order as per the enhancement. 
Please let me know, I think you need to improvices. 
Im using the database 
s104204262_db 
 the table name is

EOI
password is 210503